확장함수 변환 자료형 수정 Player.OnMoveable() 확장함수 적용

+++ Assets/GroundManager.cs
@@ -10,7 +10,7 @@ static public class GroundExtention
         return new Vector2Int(Mathf.RoundToInt(v3.x)
             , Mathf.RoundToInt(v3.z));
     }
-    static public Vector3 ToVector2Int(this Vector2Int v2Int, int y)
+    static public Vector3 ToVector2Int(this Vector2Int v2Int, float y)
     {
         return new Vector3(v2Int.x, y, v2Int.y);
     }
